General fixes
Fix code smells, Optimize imports

+++ src/main/java/it/polimi/ingsw/client/view/cli/CLIScanner.java
@@ -9,7 +9,6 @@ import it.polimi.ingsw.server.model.cards.gods.God;
 
 import java.io.InputStream;
 import java.util.*;
-import java.util.function.Consumer;
 import java.util.function.Function;
 
 public class CLIScanner<S> {
@@ -92,7 +91,7 @@ public class CLIScanner<S> {
     }
 
     private S parseStringReducedDemandCell(String string) {
-        int x = (int) (string.charAt(0) - 48);
+        int x = string.charAt(0) - 48;
         int y = string.charAt(2) - 48;
 
         return (S) (new ReducedDemandCell(x, y));
